Made it possible to send javap output to file instead of load full jar contents into memory.
Groovy jar caused test runner to run out of memory. To avoid that, for very large jars, a file is created to hold the full contents of the javap output for the jar so that only a tiny part of the output is actually kept in memory at any time.